third_party/WebKit/LayoutTests/editing/style/justify-left-crash.html

<!doctype html>
<html hidden>
<head>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
html, head, style {
    display: block;
}
</style>
</head>
<body>
<div id="sample">foo</div>
<script>
    document.designMode = 'on';
    document.execCommand('selectAll');
    document.execCommand('italic');
    document.execCommand('justifyLeft');
    document.documentElement.removeAttribute('hidden');

    test(() => {
        var sample = document.getElementById('sample');
        assert_equals(sample.innerHTML, '<i>foo</i>');
        // TODO(yosin): Once we do lazy visible position canonicalization,
        // we should have a caret.
        assert_equals(getSelection().rangeCount, 0);
    }, 'This is a crash test.');
</script>
</body>
</html>
